+++ CytoSmartOpenAPI/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""This is a python wrapper for the CytoSMART windows app to use it headless"""
+
+__author__ = """Tom Nijhof"""
+__version__ = "1.2.0"
+
+# Add here import to all the functions you need
+from .main import CytoSmartOpenAPI