+++ integration_fluo.py
@@ -2,12 +2,12 @@
 import os
 import time
 
-from luxconnector import LuxConnector
+from CytoSmartOpenAPI import CytoSmartOpenAPI
 
 result_folder = os.path.join("results", "single_image")
 os.makedirs(result_folder, exist_ok=True)
 
-connector = LuxConnector(number_of_devices=1)
+connector = CytoSmartOpenAPI(number_of_devices=1, warranty=False)
 
 serial_number = connector.get_all_serial_numbers()[0]
 
@@ -16,7 +16,14 @@ s = time.time()
 connector.set_camera_settings(serial_number, "BRIGHTFIELD", 10)
 connector.set_active_camera(serial_number, "BRIGHTFIELD")
 img = connector.get_image(serial_number)
-img.save(os.path.join(result_folder, f"{serial_number}_BRIGHTFIELD_.jpg"), "JPEG")
+img.save(os.path.join(result_folder, f"{serial_number}_BRIGHTFIELD10_.jpg"), "JPEG")
+
+# BRIGHTFIELD
+s = time.time()
+connector.set_camera_settings(serial_number, "BRIGHTFIELD", 5)
+connector.set_active_camera(serial_number, "BRIGHTFIELD")
+img = connector.get_image(serial_number)
+img.save(os.path.join(result_folder, f"{serial_number}_BRIGHTFIELD5_.jpg"), "JPEG")
 
 # RED
 s = time.time()