+++ luxconnector/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-"""This is a python wrapper around the Lux Client windows solution"""
-
-__author__ = """Tom Nijhof"""
-__version__ = "1.1.1"
-
-# Add here import to all the functions you need
-from .main import LuxConnector